Separate entities and business objects
Before there were only business objects. When I started
peristing them they became a mass and I decided to separate
persistance-level objects and business objects.

Also, testcontainers are introduced.

+++ src/main/java/com/tpakhomova/tms/service/impl/UserServiceImpl.java
@@ -1,12 +1,14 @@
 package com.tpakhomova.tms.service.impl;
 
 import com.tpakhomova.tms.data.User;
+import com.tpakhomova.tms.persistence.CommentsRepository;
+import com.tpakhomova.tms.persistence.TaskRepository;
+import com.tpakhomova.tms.persistence.data.UserEntity;
 import com.tpakhomova.tms.persistence.UserRepository;
 import com.tpakhomova.tms.service.UserService;
 import org.springframework.stereotype.Component;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.Optional;
 
 @Component
 public class UserServiceImpl implements UserService {
@@ -18,18 +20,44 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User findUser(String email) {
-        return userRepository.findByEmail(email).get();
+        Optional<UserEntity> ue = userRepository.findByEmail(email);
+
+        return ue.map(this::convertUserEntity).orElse(null);
     }
 
     @Override
     public boolean createUser(User user) {
-        userRepository.save(user);
+        userRepository.save(convertUser(user));
+        userRepository.flush();
         return true;
     }
 
     @Override
     public boolean deleteUser(String email) {
+        if (userRepository.findByEmail(email).isEmpty()) {
+            return false;
+        }
+
         userRepository.deleteByEmail(email);
         return true;
+    }
+
+    private User convertUserEntity(UserEntity ue) {
+        return new User(
+                ue.getId(), ue.getUsername(), ue.getPassHash(), ue.getEmail(),
+                ue.getFirstName(), ue.getLastName()
+        );
+    }
+
+    private UserEntity convertUser(User user) {
+        var ue = new UserEntity();
+        ue.setId(user.getId());
+        ue.setUsername(user.getUsername());
+        ue.setEmail(user.getEmail());
+        ue.setFirstName(user.getFirstName());
+        ue.setLastName(user.getLastName());
+        ue.setPassHash(user.getPassHash());
+
+        return ue;
     }
 }
